pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tianxing"
version = "0.1.0"
description = "Automated paper review and improvement tools for Claude Code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
tianxing = "tianxing.__init__:main"

[tool.setuptools.packages.find]
include = ["tianxing*"]
